[EX-126] [M2] Send historical orders to Orders API - applied logic to run sync process only store by store instead of global or website scopes - added message about number of order sync process
