+++ Block/System/Config/Orders/Button.php
@@ -11,19 +11,13 @@
 namespace Extend\Warranty\Block\System\Config\Orders;
 
 use Exception;
-use Extend\Warranty\Helper\Api\Data as DataHelper;
-use Magento\Backend\Block\Template\Context;
-use Magento\Config\Block\System\Config\Form\Field;
-use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Framework\Data\Form\Element\AbstractElement;
-use Magento\Framework\Stdlib\DateTime\TimezoneInterface;
+use Extend\Warranty\Block\System\Config\AbstractSyncButton;
 use Magento\Store\Model\ScopeInterface;
-use Magento\Store\Model\Store;
 
 /**
  * Class Button
  */
-class Button extends Field
+class Button extends AbstractSyncButton
 {
     /**
      * Path to template file in theme
@@ -31,67 +25,6 @@ class Button extends Field
      * @var string
      */
     protected $_template = "Extend_Warranty::system/config/orders/button.phtml";
-
-    /**
-     * @var TimezoneInterface
-     */
-    private $timezone;
-
-    /**
-     * Data Helper
-     *
-     * @var DataHelper
-     */
-    private $dataHelper;
-
-    /**
-     * Button constructor
-     *
-     * @param Context $context
-     * @param TimezoneInterface $timezone
-     * @param DataHelper $dataHelper
-     * @param array $data
-     */
-    public function __construct(
-        Context $context,
-        TimezoneInterface $timezone,
-        DataHelper $dataHelper,
-        array $data = []
-    ) {
-        $this->timezone = $timezone;
-        $this->dataHelper = $dataHelper;
-        parent::__construct($context, $data);
-    }
-
-    /**
-     * Remove scope label
-     *
-     * @param  AbstractElement $element
-     * @return string
-     */
-    public function render(AbstractElement $element): string
-    {
-        $element->unsScope();
-        $element->unsCanUseWebsiteValue();
-        $element->unsCanUseDefaultValue();
-
-        return parent::render($element);
-    }
-
-    /**
-     * Render value
-     *
-     * @param AbstractElement $element
-     * @return string
-     */
-    protected function _renderValue(AbstractElement $element): string
-    {
-        $html = '<td class="value">';
-        $html .= $this->toHtml();
-        $html .= '</td>';
-
-        return $html;
-    }
 
     /**
      * Get last sync date
@@ -109,43 +42,16 @@ class Button extends Field
     }
 
     /**
-     * Get scope data
-     *
      * @return array
+     * @throws \Magento\Framework\Exception\LocalizedException
      */
-    public function getScopeData(): array
+    public function getSyncUrls()
     {
-        $request = $this->getRequest();
-        $website = $request->getParam('website');
-        $store = $request->getParam('store');
-
-        if ($website) {
-            $scopeType = ScopeInterface::SCOPE_WEBSITE;
-            $scopeId = $website;
-        } elseif ($store) {
-            $scopeType = ScopeInterface::SCOPE_STORE;
-            $scopeId = $store;
-        } else {
-            $scopeType = ScopeConfigInterface::SCOPE_TYPE_DEFAULT;
-            $scopeId = Store::DEFAULT_STORE_ID;
+        $storeIds = $this->getScopeStoreIds();
+        $syncUrls = [];
+        foreach ($storeIds as $storeId) {
+            $syncUrls[] = $this->getUrl('extend/orders/historicalorders', [ScopeInterface::SCOPE_STORE => $storeId]);
         }
-
-        return [
-            'scopeType' => $scopeType,
-            'scopeId'   => $scopeId,
-        ];
-    }
-
-    /**
-     * Check if button enabled
-     *
-     * @return bool
-     */
-    public function isButtonEnabled(): bool
-    {
-        $scopeData = $this->getScopeData();
-
-        return $this->dataHelper->getStoreId($scopeData['scopeType'], $scopeData['scopeId'])
-            && $this->dataHelper->getApiKey($scopeData['scopeType'], $scopeData['scopeId']);
+        return $syncUrls;
     }
 }
